♻️ refactor : useImageDownloader 훅 경로 수정 및 불필요한 훅 삭제 #386
